Almost entirely useless commit
Worked a lot, broke it so bad I rolled it back. That just broke it
further, worked on that a lot, deleted and re-cloned from last commit on
github.  Managed to save some of the changes but nothing major.  Which
is good because like I said I broke it really bad.